+++ src/slog_test.rs
@@ -51,6 +51,7 @@
 //! For verifying statistics, the `create_logger_buffer` and `check_expected_stats` methods
 //! are useful for creating a `slog_extlog::StatisticsLogger` and then verifying that statistics
 //! are generated as expected.
+#![allow(clippy::float_cmp)]
 pub extern crate erased_serde;
 pub extern crate iobuffer;
 pub extern crate serde_json;
@@ -231,7 +232,8 @@ pub fn check_expected_stat_snaphots(
         match found_stat.values {
             StatSnapshotValues::Counter(ref vals) | StatSnapshotValues::Gauge(ref vals) => {
                 for value in &stat.values {
-                    let found_value = vals.iter()
+                    let found_value = vals
+                        .iter()
                         .find(|val| val.group_values == value.group_values);
 
                     assert!(